tests/submission/Microsoft/forms/formmethod.html

<!DOCTYPE HTML>
<html>
    <head>
        <title>formmethod on button element</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta name="description" content="formmethod" />
    </head>
    <body>
        <script src="/resources/testharness.js" ></script>
        <script src="/resources/testharnessreport.js" ></script>

        <div id=log></div>
        <button style="display: none" type="submit" formmethod="post">Submit</button>
        <input style="display: none" type="submit" formmethod="post" value="submit"></input>
        <input style="display: none" type="submit" formmethod="get" value="submit"></input>
        
        
        <script type="text/javascript">
        
            test(function() {assert_true("formMethod" in document.getElementsByTagName("button")[0])}, "formMethod attribute 'in' support fon button element");
            test(function() {assert_equals(document.getElementsByTagName("button")[0].formMethod.toLowerCase(), "post")}, "formMethod returns post for button element");

            test(function() {assert_true("formMethod" in document.getElementsByTagName("input")[0])}, "formMethod attribute 'in' support on input element");
            test(function() {assert_equals(document.getElementsByTagName("input")[0].formMethod.toLowerCase(), "post")}, "formMethod returns post for input element");

            var testElem = document.getElementsByTagName("input")[1];
            testElem.formMethod = "post";
            
            test(function() {assert_true("formMethod" in document.getElementsByTagName("input")[1])}, "formMethod attribute 'in' support on input element exsists after modification");
            test(function() {assert_equals(document.getElementsByTagName("input")[1].formMethod.toLowerCase(), "post")}, "formMethod returns correct value after being changed by the DOM");
            
        </script>
    </body>
</html>
